added spinner on the shoe details page and edit shoe form when fetching data

+++ server/src/services/cartService.js
@@ -66,9 +66,7 @@ const cartService = {
     },
 
     async deleteCart(userId) {
-        await Cart.findOneAndDelete(
-            { owner: userId }
-        )
+        await Cart.findOneAndDelete({ owner: userId })
     }
 
 };
